cleanup

+++ interfaces/src/main/java/online/ronikier/todo/interfaces/web/TaskController.java
@@ -13,6 +13,7 @@ import online.ronikier.todo.infrastructure.service.TaskService;
 import online.ronikier.todo.interfaces.mappers.TaskMapper;
 import online.ronikier.todo.library.Parameters;
 import online.ronikier.todo.library.Utilities;
+import online.ronikier.todo.templete.FormAction;
 import online.ronikier.todo.templete.SuperController;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.PropertySource;
@@ -71,7 +72,7 @@ public class TaskController extends SuperController {
         if (!securityCheckOK(model)) return "login";
         refreshForm(taskForm, model);
         if (taskForm.getTask().getId() == null) {
-            taskForm.setTask(initializeTask());
+            taskForm.setTask(taskService.initializeTask());
         }
         return "task";
     }
@@ -109,32 +110,33 @@ public class TaskController extends SuperController {
 
         if (!securityCheckOK(model)) return "/";
 
+        if (taskForm.getAction() == null) return "task";
+
         Task processedTask;
 
         //TODO: Refactor - Transfer to service
         switch (taskForm.getAction()) {
-            case "delete":
+            case DELETE:
                 processedTask = taskService.findTaskById(taskForm.getTaskId()).get();
                 taskService.deleteTaskById(processedTask.getId());
                 initializeForm(taskForm,model);
                 break;
-            case "reject":
+            case REJECT:
                 processedTask = taskService.findTaskById(taskForm.getTaskId()).get();
                 processedTask.setTaskState(TaskState.REJECTED);
                 processedTask.setDue(null);
-                processedTask.setTaskStatus(TaskStatus.UNKNOWN);
                 taskService.saveTask(processedTask);
                 break;
-            case "filter":
+            case FILTER:
                 break;
-            case "complete": {
+            case COMPLETE: {
                 processedTask = taskService.findTaskById(taskForm.getTaskId()).get();
                 processedTask.setTaskState(TaskState.COMPLETED);
                 processedTask.setTaskStatus(TaskStatus.OK);
                 taskService.saveTask(processedTask);
                 break;
             }
-            case "save": {
+            case SAVE: {
                 if (bindingResult.hasErrors()) {
                     log.error(Messages.ERROR_TASK_ADD);
                     bindingResult.getAllErrors().forEach(error -> log.error(Messages.SEPARATOR + error.toString()));
@@ -142,20 +144,21 @@ public class TaskController extends SuperController {
                     break;
                 }
                 try {
-                    processedTask = taskService.findTaskByName(taskForm.getName());
+                    Optional<Task> optionalTask = taskService.findTaskByName(taskForm.getName());
                     //processedTask = Parameters.SYSTEM_ALLOW_DUPLICATE_NAMES
                     //        ? taskService.findTaskByName(taskForm.getName())
                     // TODO: Fix null pointer by new task - introduce New button (visible/disabled)
                     //        : taskService.findTaskById(taskForm.getTaskId()).get();
-                    if (processedTask != null) {
-                        log.info((Messages.INFO_TASK_EXISTS));
+                    if (optionalTask.isPresent()) {
+                        log.info((Messages.TASK_EXISTS));
+                        processedTask = optionalTask.get();
                         processedTask.setTaskState(TaskState.MODIFIED);
                         log.debug(processedTask.toString());
                     } else {
-                        processedTask = initializeTask();
+                        processedTask = taskService.initializeTask();
                     }
                     saveTask(taskForm, processedTask);
-                    log.info(Messages.INFO_TASK_CREATED);
+                    log.info(Messages.TASK_CREATED);
                     log.debug(processedTask.toString());
                 } catch (Exception e) {
                     taskForm.setShowDialog(true);
@@ -166,7 +169,7 @@ public class TaskController extends SuperController {
         }
         refreshForm(taskForm, model);
         if (taskForm.getTask() == null) {
-            taskForm.setTask(initializeTask());
+            taskForm.setTask(taskService.initializeTask());
         }
         return "task";
     }
@@ -182,7 +185,7 @@ public class TaskController extends SuperController {
 
         if (!securityCheckOK(model)) return "/";
 
-        log.info(Messages.INFO_TASK_DELETING);
+        log.info(Messages.TASK_DELETING);
         taskService.deleteTaskById(taskId);
         initializeForm(taskForm, model);
         model.addAttribute("tasksRequiredTasks", null);
@@ -194,7 +197,7 @@ public class TaskController extends SuperController {
      * @param task
      */
     private void saveTask(TaskForm taskForm, Task task) {
-        log.info(Messages.INFO_TASK_MODIFIED);
+        log.info(Messages.TASK_MODIFIED);
         taskMapper.form2Domain(taskForm, task);
 
         assignResponsible(taskForm, task);
@@ -212,7 +215,7 @@ public class TaskController extends SuperController {
             if (Parameters.SYSTEM_SAVE_NOT_REQUIRED) {
                 taskService.saveTask(task);
             } else {
-                log.info(Messages.INFO_SKIPPING_NOT_REQUIRED);
+                log.info(Messages.SKIPPING_NOT_REQUIRED);
                 return;
             }
         }
@@ -220,7 +223,7 @@ public class TaskController extends SuperController {
         taskForm.getTask().setTaskType(task.getTaskType());
         taskForm.getTask().setTaskState(task.getTaskState());
 
-        log.info(Messages.INFO_TASK_MODIFIED);
+        log.info(Messages.TASK_MODIFIED);
         log.debug(task.toString());
     }
 
@@ -237,49 +240,9 @@ public class TaskController extends SuperController {
     }
 
 
-    /**
-     * @param taskForm
-     * @return
-     * @throws ParseException
-     */
-    private Task initializeTask() {
-
-        //TODO: Implement system common tasks
-        //return new Task(dafaultTasks,
 
 
-        Task newTask = new Task(
-                appendMaintanceTasks(),
-                null,
-                null,
-                null,
-                Utilities.dateCurrent(),
-                null,       //Utilities.dateCurrent(),
-                null,       //Utilities.dateFuture(taskCompletionTimeDays),
-                null,
-                null,
-                Double.valueOf(1),
-                CostUnit.SOLDIER,
-                TaskState.NEW,
-                TaskType.GENERAL,
-                TaskStatus.OK
-        );
 
-        return newTask;
-
-    }
-
-    /**
-     * @param taskName
-     * @return
-     */
-    private List<Task> appendMaintanceTasks() {
-        if (Parameters.SYSTEM_SKIP_MAINTENANCE_TASKS) {
-            log.info(Messages.INFO_SKIPPING_MAINTENANCE_TASKS);
-            return null;
-        }
-        return taskService.getMaintanceTasks();
-    }
 
     /**
      * @param taskId
@@ -311,7 +274,7 @@ public class TaskController extends SuperController {
      * @param model
      */
     private void initializeForm(TaskForm taskForm, Model model) {
-        taskForm.setTask(initializeTask());
+        taskForm.setTask(taskService.initializeTask());
         taskForm.setTaskFilterForm(new TaskFilterForm());
         //taskForm.setShowTaskDetails(false);
         refreshForm(taskForm, model);
